i965: Push UBO data via 3DSTATE_CONSTANT_XS (but don't use it yet).
With this patch, we actually begin uploading UBO data as push constants,
using the ranges we chose in the previous commit.  However, it does not
yet update the shader to *use* this newly pushed data - it will still
pull the data again, rather than reading the pushed values.

This requires three changes:
- Update 3DSTATE_CONSTANT_XS to actually use the UBO ranges.
- Update 3DSTATE_PS/WM bits to signal that there are push constants
  to be uploaded, even if there are no regular uniforms.
- Update the compiler to skip over the additional push constant
  registers, so that inputs continue working.

Signed-off-by: Kenneth Graunke <kenneth@whitecape.org>