i965: Delete BRW_NEW_PUSH_CONSTANT_ALLOCATION.
Previously we would re-upload the constant data to the batchbuffer,
then re-emit the packets.  We only need to do the last step (causing
the existing data in the batchbuffer to be re-uploaded to the push
constant staging area in the L3).

Now that we've separated the two, it's pretty easy to accomplish.

Signed-off-by: Kenneth Graunke <kenneth@whitecape.org>