egl: wayland: bufferAge: only request a back buffer if there isn't one
Seeing gtk+ application lockup when they query the buffer age of a surface.

Since we update the buffer age field only when creating buffers & swaping
them on the client side, there shouldn't be any need for requesting a new
back buffer if there is already one available.

This change depends on a previous commit settings the age value of buffers
to 0 on buffer destruction, so we're sure to always have an up to date
value.

Signed-off-by: Lionel Landwerlin <lionel.g.landwerlin@intel.com>
Bugzilla: https://bugs.freedesktop.org/show_bug.cgi?id=84016