intel: add aubinator ui
The motivation for building this UI is to help debug the intel
drivers. It can be tricky to figure what is going wrong in the
emission of commands (figuring out if there is an off by one error in
the sampler state pointers for example or if you accidently override
some bits of memory). Going through the text output of aubinator can
be quite tedious (time consuming and error prone), as you need to go
back and forth several pages of dumps to build a mental model of the
state of the GPU for a given draw call. This calls for a better way to
display/search an aubdump.

This UI is mostly based off the original aubinator code. It's using
the ImGui toolkit [1]. Being an immediate mode UI toolkit, it makes
the code of the original aubinator pretty easy to reuse/port to
generate a UI.

This also adds an interesting feature which collects the GPU state by
walking the instruction stream and "snapshotting" the last sets of
commands allowing to present what would be the state of the GPU at a
given point (right now it only snapshots on 3DPRIMITIVE commands).

Some screenshots :
    https://i.imgur.com/0JTLkTo.png
    https://i.imgur.com/ABq31XD.png

Although the UI toolkit isn't tied to any of the main toolkit (like
Gtk+, Qt, etc...), we still require a connection to the desktop
display. The ImGui backend currently used is built on Gtk+ & Cogl. The
choice of Gtk+ is mostly due to having UI scaling working properly on
wayland/x11 (on hidpi screens). Cogl mostly because it's more
convenient that raw GL. Of course this could be revisited.

[1] : https://github.com/ocornut/imgui

Signed-off-by: Lionel Landwerlin <lionel.g.landwerlin@intel.com>

For the autotools part:
Reviewed-by: Emil Velikov <emil.velikov@collabora.com>