meson: don't us build_by_default for specific gallium drivers
Using build_by_default : false is convenient for dependencies that can
be pulled in by various diverse components of the build system, the
gallium hardware/software drivers and state trackers do not fit that
description. Instead, these should be guarded using the variable that tracks
whether that driver should be enabled.

This leaves a few helper libraries: trace, rbug, etc. as
build_by_default : false because there are a large number of gallium
components that pull them in. This fixes odd build errors since the
default drivers patch landed.

Fixes: 1873327c4b5a09 ("meson: implement default driver arguments")
Signed-off-by: Dylan Baker <dylanx.c.baker@intel.com>