intel/aub_write: init context images with empty batchbuffer
When we create a context image, we allocate some memory and write a
few commands in it to bootstrap the engine. At the end of the first
run of this context, HW will write default values back for lots of
register we didn't setup. This is how i915 gets its default image.

Here we apply the same trick so that the first time we execute a batch
buffer, the default values are already in the image.

Signed-off-by: Lionel Landwerlin <lionel.g.landwerlin@intel.com>