intel/aub_mem: use larger physical page size
With error states reported by Jason on DOOM, we're running out of
virtual mapping in the process inspecting the aub file. This is
because we require +4Gb of physical memory (which is fine on a machine
with decent amount of RAM) which is more than 65K mappings of 4k pages
and Linux has a default limit of 65K (/proc/sys/vm/max_map_count).

We can easilly workaround this issue by having larger mappings for the
physical pages (one mapping containing more than one page). Here we
select 32Kb.

Signed-off-by: Lionel Landwerlin <lionel.g.landwerlin@intel.com>