intel: limit urb size for SKL GT1
The documentation puts the URB size for SKL GT1 as "128-192K". I guess
this means we can't guess which one it is, so we have to go for the
lower bound. This change reuses values from Gen9LP (since it has the
same 128K URB size). I'm not completely sure this is correct at least
it fixes a CTS test :

  dEQP-GLES31.functional.geometry_shading.layered.render_with_default_layer_3d

Signed-off-by: Lionel Landwerlin <lionel.g.landwerlin@intel.com>
Bugzilla: https://bugs.freedesktop.org/show_bug.cgi?id=107505