anv: use format stored aspects to build up images & imageviews
For single planar YCbCr formats like :

   - VK_FORMAT_G8B8G8R8_422_UNORM
   - VK_FORMAT_B8G8R8G8_422_UNORM
   - VK_FORMAT_G16B16G16R16_422_UNORM
   - VK_FORMAT_B16G16R16G16_422_UNORM

we could only use YCbCr HW formats to sample from those. This is
because the HW was designed in such a way that accessing a same memory
region with 2 different views would make the HW samplers yield
incorrect results. This restriction got removed in Gen11/ICL and we
are now considering sampling from those formats using 2 entries in the
binding/sampler tables.

This change allows a same buffer to be bound to several planes,
allowing multiple views of the same buffer by looking at the aspect on
the anv format description.

Signed-off-by: Lionel Landwerlin <lionel.g.landwerlin@intel.com>