intel: add gpu dbus service
The current role of this dbus service is to handle privileged
operations on the device. There are 2 kinds to privileged operations
we want mediated with regard to performance analysis :

  - Access to performance hardware can expose how other applications
    are using the GPU, this currently requires a special setting in
    the i915 driver to allow non privileged applications to access that
    capability (sudo sysctl dev.i915.perf_stream_paranoid=0).

  - Access to setting GPU min/max/boost frequencies, this is required
    by some performance analysis applications. Setting a particular
    (usually max) frequency helps having more acurate performance
    data.

This service is conditional to the intel-service option in meson and
disabled by default for now. When enabled, a new DBus service is
installed and the i965 driver will use (in the following patch) dbus
to access those privileged operations. When not enabled we fallback to
the currently behavior of the driver which is to try to use privileged
operations fail accordingly if not available when trying to use the
GL_INTEL_performance_query extension.

This service is not built with autotools.

Signed-off-by: Lionel Landwerlin <lionel.g.landwerlin@intel.com>