i965: request display usage for images created with modifiers
The GBM API is missing a usage parameter for
gbm_bo_create_with_modifiers(). This is causing problems for
application that want to flip a buffer on KMS because our HW's
alignement constraints for display are more restrictive than for
render target. To address this issue, assume that if a modifier is
ever specified the display usage should be considered.

Signed-off-by: Lionel Landwerlin <lionel.g.landwerlin@intel.com>
Cc: <mesa-stable@lists.freedesktop.org>
Closes: #1825